Added new events and added them to the boundaries
Added MrJ and Crystal events, which trigger text above the player. Also updated the boundaries of the map to give the player a reason why they should remain inside.

Updated the materials for several of the buildings, and next step is to randomzie their colors a bit to make it more intersting.